requirements.txt
# Django
Django==3.2.7

# Guicorn
gunicorn==20.1.0

# Database Postgres
psycopg2-binary==2.9.1

# Django Rest Framework
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-nested-routers==0.93.3
django-filter==21.1
django-rest-auth==0.9.5
django-allauth==0.45.0

# Admin
django-admin-interface==0.16.3
django-ckeditor==6.1.0

# Project Dependencies
Pillow==8.3.2

# Envs
django-dotenv==1.4.2

# Heroku
django-heroku==0.0.0